refactor: split up the core volume tests into modules
