+++ control-plane/agents/core/tests/volume/mod.rs
@@ -1,41 +1,31 @@
 #![cfg(test)]
 
+mod capacity;
+mod garbage_collection;
+mod helpers;
+mod hotspare;
+
 use common_lib::{
     store::etcd::Etcd,
     transport_api::{v0::Replicas, ReplyError, ReplyErrorKind, ResourceKind},
     types::v0::{
-        openapi::apis::{StatusCode, Uuid},
+        openapi::{models, models::NodeStatus},
         store::{
             definitions::Store,
             nexus_persistence::{NexusInfo, NexusInfoKey},
         },
         transport::{
-            Child, ChildState, CreateReplica, CreateVolume, DestroyVolume, Filter, GetNexuses,
-            GetReplicas, GetVolumes, Nexus, NodeId, PublishVolume, SetVolumeReplica, ShareVolume,
-            Topology, UnpublishVolume, UnshareVolume, Volume, VolumeShareProtocol, VolumeState,
+            Child, ChildState, CreateVolume, DestroyVolume, Filter, GetNexuses, GetReplicas,
+            GetVolumes, Nexus, NodeId, PublishVolume, SetVolumeReplica, ShareVolume, Topology,
+            UnpublishVolume, UnshareVolume, Volume, VolumeId, VolumeShareProtocol, VolumeState,
             VolumeStatus,
         },
     },
 };
-
 use deployer_cluster::{Cluster, ClusterBuilder};
-use rpc::io_engine::FaultNexusChildRequest;
-
-use common_lib::{
-    transport_api::TimeoutOptions,
-    types::v0::{
-        openapi::{models, models::NodeStatus, tower::client::Error},
-        store::{definitions::StorableObject, volume::VolumeSpec},
-        transport::{
-            ChildUri, CreateNexus, DestroyReplica, GetSpecs, NexusId, ReplicaId, ReplicaOwners,
-            VolumeId,
-        },
-    },
-};
 use grpc::operations::{
     nexus::traits::NexusOperations, node::traits::NodeOperations,
-    registry::traits::RegistryOperations, replica::traits::ReplicaOperations,
-    volume::traits::VolumeOperations,
+    replica::traits::ReplicaOperations, volume::traits::VolumeOperations,
 };
 use std::{
     convert::{TryFrom, TryInto},
@@ -73,1093 +63,6 @@ async fn test_volume(cluster: &Cluster) {
 }
 
 const RECONCILE_TIMEOUT_SECS: u64 = 7;
-
-#[tokio::test]
-async fn hotspare() {
-    let cluster = ClusterBuilder::builder()
-        .with_rest(true)
-        .with_agents(vec!["core"])
-        .with_io_engines(3)
-        .with_pools(2)
-        .with_cache_period("1s")
-        .with_reconcile_period(Duration::from_secs(1), Duration::from_secs(1))
-        .build()
-        .await
-        .unwrap();
-
-    let node_client = cluster.grpc_client().node();
-    let nodes = node_client.get(Filter::None, None).await.unwrap();
-    tracing::info!("Nodes: {:?}", nodes);
-
-    hotspare_faulty_children(&cluster).await;
-    hotspare_unknown_children(&cluster).await;
-    hotspare_missing_children(&cluster).await;
-    hotspare_replica_count(&cluster).await;
-    hotspare_replica_count_spread(&cluster).await;
-    hotspare_nexus_replica_count(&cluster).await;
-}
-
-const POOL_SIZE_BYTES: u64 = 128 * 1024 * 1024;
-#[tokio::test]
-async fn volume_nexus_reconcile() {
-    let cluster = ClusterBuilder::builder()
-        .with_rest(true)
-        .with_agents(vec!["core"])
-        .with_io_engines(2)
-        .with_tmpfs_pool(POOL_SIZE_BYTES)
-        .with_cache_period("1s")
-        .with_reconcile_period(Duration::from_secs(1), Duration::from_secs(1))
-        .build()
-        .await
-        .unwrap();
-
-    let node_client = cluster.grpc_client().node();
-    let nodes = node_client.get(Filter::None, None).await.unwrap();
-    tracing::info!("Nodes: {:?}", nodes);
-
-    missing_nexus_reconcile(&cluster).await;
-}
-
-#[tokio::test]
-async fn garbage_collection() {
-    let reconcile_period = Duration::from_millis(500);
-    let cluster = ClusterBuilder::builder()
-        .with_rest(true)
-        .with_agents(vec!["core"])
-        .with_io_engines(3)
-        .with_tmpfs_pool(POOL_SIZE_BYTES)
-        .with_cache_period("1s")
-        .with_reconcile_period(reconcile_period, reconcile_period)
-        .build()
-        .await
-        .unwrap();
-
-    let node_client = cluster.grpc_client().node();
-    let nodes = node_client.get(Filter::None, None).await.unwrap();
-    tracing::info!("Nodes: {:?}", nodes);
-
-    unused_nexus_reconcile(&cluster).await;
-    unused_reconcile(&cluster).await;
-    deleting_volume_reconcile(&cluster).await;
-    offline_replicas_reconcile(&cluster, reconcile_period).await;
-}
-
-async fn deleting_volume_reconcile(cluster: &Cluster) {
-    let client = cluster.grpc_client().volume();
-    let volume = client
-        .create(
-            &CreateVolume {
-                uuid: "1e3cf927-80c2-47a8-adf0-95c486bdd7b7".try_into().unwrap(),
-                size: 5242880,
-                replicas: 1,
-                ..Default::default()
-            },
-            None,
-        )
-        .await
-        .unwrap();
-
-    client
-        .publish(
-            &PublishVolume {
-                uuid: volume.uuid().clone(),
-                target_node: None,
-                share: None,
-            },
-            None,
-        )
-        .await
-        .unwrap();
-
-    // 1. Pause etcd
-    cluster.composer().pause("etcd").await.unwrap();
-
-    // 2. Attempt to delete the volume
-    client
-        .destroy(
-            &DestroyVolume {
-                uuid: volume.uuid().to_owned(),
-            },
-            None,
-        )
-        .await
-        .expect_err("ETCD is paused...");
-
-    // 3. Bring back etcd
-    cluster.composer().thaw("etcd").await.unwrap();
-
-    // 4. Wait for volume deletion
-    wait_till_volume_deleted(cluster).await;
-
-    // 5. Volume replicas and nexuses should have been deleted as well
-    let specs = cluster.rest_v00().specs_api().get_specs().await.unwrap();
-    assert!(specs.nexuses.is_empty());
-    let nexuses = cluster.rest_v00().nexuses_api().get_nexuses().await;
-    assert!(nexuses.unwrap().is_empty());
-    assert!(specs.replicas.is_empty());
-    let replicas = cluster.rest_v00().replicas_api().get_replicas().await;
-    assert!(replicas.unwrap().is_empty());
-}
-
-/// Wait for a volume to reach the provided status
-async fn wait_till_volume_deleted(cluster: &Cluster) {
-    let timeout = Duration::from_secs(RECONCILE_TIMEOUT_SECS);
-    let client = cluster.grpc_client().volume();
-    let start = std::time::Instant::now();
-    loop {
-        let volumes = client.get(Filter::None, None, None).await.unwrap();
-        if volumes.entries.is_empty() {
-            return;
-        }
-
-        if std::time::Instant::now() > (start + timeout) {
-            panic!(
-                "Timeout waiting for the volumes to be deleted current: '{:?}'",
-                volumes
-            );
-        }
-        tokio::time::sleep(Duration::from_millis(500)).await;
-    }
-}
-
-async fn offline_replicas_reconcile(cluster: &Cluster, reconcile_period: Duration) {
-    let rest_api = cluster.rest_v00();
-    let volumes_api = rest_api.volumes_api();
-
-    let volume = volumes_api
-        .put_volume(
-            &"1e3cf927-80c2-47a8-adf0-95c481bdd7b7".parse().unwrap(),
-            models::CreateVolumeBody::new(models::VolumePolicy::default(), 2, 5242880u64, false),
-        )
-        .await
-        .unwrap();
-
-    let nodes = rest_api.nodes_api().get_nodes().await.unwrap();
-    let replica_nodes = rest_api.replicas_api().get_replicas().await.unwrap();
-    let replica_nodes = replica_nodes
-        .into_iter()
-        .map(|r| r.node)
-        .collect::<Vec<_>>();
-    let free_node = nodes
-        .into_iter()
-        .find_map(|n| {
-            if replica_nodes.iter().all(|repl_node| repl_node != &n.id) {
-                Some(n.id)
-            } else {
-                None
-            }
-        })
-        .unwrap();
-
-    // 1. publish on the node with no replicas
-    let volume = volumes_api
-        .put_volume_target(
-            &volume.spec.uuid,
-            &free_node,
-            models::VolumeShareProtocol::Nvmf,
-        )
-        .await
-        .unwrap();
-
-    tracing::info!("Volume: {:?}", volume);
-    let volume_state = volume.state;
-
-    let volume_id = volume_state.uuid;
-    let volume = volumes_api.get_volume(&volume_id).await.unwrap();
-    assert_eq!(volume.state.status, models::VolumeStatus::Online);
-
-    // 2. kill all replica nodes
-    for node in replica_nodes.iter().filter(|n| n != &&free_node) {
-        cluster.composer().stop(node).await.unwrap();
-    }
-
-    // 3. wait for volume to become Faulted
-    wait_till_volume_status(cluster, &volume.spec.uuid, models::VolumeStatus::Faulted).await;
-
-    // 4. restart the core-agent
-    cluster.restart_core().await;
-
-    cluster.volume_service_liveness(None).await.unwrap();
-    wait_till_volume_status(cluster, &volume.spec.uuid, models::VolumeStatus::Faulted).await;
-
-    // 5. After the reconcilers run, replicas should not have been disowned
-    tokio::time::sleep(reconcile_period * 3).await;
-
-    let volume = volumes_api.get_volume(&volume.spec.uuid).await.unwrap();
-    assert_eq!(volume.state.status, models::VolumeStatus::Faulted);
-
-    let replicas = rest_api.specs_api().get_specs().await.unwrap().replicas;
-    assert_eq!(replicas.len(), 2);
-    assert_eq!(
-        replicas
-            .iter()
-            .filter(|r| r.owners.volume.as_ref() == Some(&volume.spec.uuid))
-            .count(),
-        2
-    );
-
-    volumes_api.del_volume(&volume_id).await.unwrap();
-}
-
-async fn unused_nexus_reconcile(cluster: &Cluster) {
-    let rest_api = cluster.rest_v00();
-    let volumes_api = rest_api.volumes_api();
-    let nexus_client = cluster.grpc_client().nexus();
-
-    let volume = volumes_api
-        .put_volume(
-            &"1e3cf927-80c2-47a8-adf0-95c481bdd7b7".parse().unwrap(),
-            models::CreateVolumeBody::new(models::VolumePolicy::default(), 2, 5242880u64, false),
-        )
-        .await
-        .unwrap();
-
-    let volume = volumes_api
-        .put_volume_target(
-            &volume.spec.uuid,
-            cluster.node(0).as_str(),
-            models::VolumeShareProtocol::Nvmf,
-        )
-        .await
-        .unwrap();
-
-    tracing::info!("Volume: {:?}", volume);
-    let volume_state = volume.state;
-
-    let volume_id = volume_state.uuid;
-    let volume = volumes_api.get_volume(&volume_id).await.unwrap();
-    assert_eq!(volume.state.status, models::VolumeStatus::Online);
-
-    let mut create_nexus = CreateNexus {
-        node: cluster.node(0),
-        uuid: NexusId::new(),
-        size: volume.spec.size,
-        children: vec![
-            "malloc:///test?size_mb=10&uuid=b9558b8c-cb22-47f3-b33b-583db25b5a8c".into(),
-        ],
-        managed: true,
-        owner: None,
-        config: None,
-    };
-    let nexus = nexus_client.create(&create_nexus, None).await.unwrap();
-    let nexus = wait_till_nexus_state(cluster, &nexus.uuid, None).await;
-    assert_eq!(nexus, None, "nexus should be gone");
-
-    create_nexus.owner = Some(VolumeId::new());
-    let nexus = nexus_client.create(&create_nexus, None).await.unwrap();
-    let nexus = wait_till_nexus_state(cluster, &nexus.uuid, None).await;
-    assert_eq!(nexus, None, "nexus should be gone");
-
-    volumes_api.del_volume(&volume_id).await.unwrap();
-}
-
-async fn unused_reconcile(cluster: &Cluster) {
-    let rest_api = cluster.rest_v00();
-    let volumes_api = rest_api.volumes_api();
-
-    let volume = volumes_api
-        .put_volume(
-            &"22054b1f-cf32-46dc-90ff-d6a5c61429c2".parse().unwrap(),
-            models::CreateVolumeBody::new(models::VolumePolicy::new(true), 2, 5242880u64, false),
-        )
-        .await
-        .unwrap();
-
-    let data_replicas_nodes = volume
-        .state
-        .replica_topology
-        .values()
-        .map(|r| r.node.clone().unwrap())
-        .collect::<Vec<_>>();
-    let nodes = rest_api.nodes_api().get_nodes().await.unwrap();
-    let unused_node = nodes
-        .iter()
-        .find(|r| !data_replicas_nodes.contains(&r.id))
-        .cloned()
-        .unwrap();
-    let nexus_node = nodes
-        .iter()
-        .find(|n| n.id != unused_node.id)
-        .cloned()
-        .unwrap();
-    let replica_nexus = volume
-        .state
-        .replica_topology
-        .into_iter()
-        .find_map(|(i, r)| {
-            if r.node.as_ref().unwrap() == &nexus_node.id {
-                Some(i)
-            } else {
-                None
-            }
-        })
-        .unwrap();
-
-    let volume = volumes_api
-        .put_volume_target(
-            &volume.spec.uuid,
-            nexus_node.id.as_str(),
-            models::VolumeShareProtocol::Nvmf,
-        )
-        .await
-        .unwrap();
-    tracing::info!("Volume: {:?}\nUnused Node: {}", volume, unused_node.id);
-
-    // 1. first we kill the node where the nexus is running
-    cluster.composer().kill(&nexus_node.id).await.unwrap();
-    // 2. now we force unpublish the volume
-    volumes_api
-        .del_volume_target(&volume.spec.uuid, Some(true))
-        .await
-        .unwrap();
-    // 3. publish on the previously unused node
-    let volume = volumes_api
-        .put_volume_target(
-            &volume.spec.uuid,
-            unused_node.id.as_str(),
-            models::VolumeShareProtocol::Nvmf,
-        )
-        .await
-        .unwrap();
-    tracing::info!("Volume: {:?}", volume);
-
-    // 4. now wait till the "broken" replica is disowned
-    wait_till_replica_disowned(cluster, replica_nexus.parse().unwrap()).await;
-
-    // 5. now wait till the volume becomes online again
-    // (because we'll add a replica a rebuild)
-    wait_till_volume_status(cluster, &volume.spec.uuid, models::VolumeStatus::Online).await;
-
-    // 6. Bring back the io-engine and the original nexus and replica should be deleted
-    cluster.composer().start(&nexus_node.id).await.unwrap();
-    let timeout = Duration::from_secs(RECONCILE_TIMEOUT_SECS);
-    let start = std::time::Instant::now();
-    loop {
-        let specs = cluster.rest_v00().specs_api().get_specs().await.unwrap();
-        if specs.nexuses.len() == 1 && specs.replicas.len() == 2 {
-            break;
-        }
-
-        if std::time::Instant::now() > (start + timeout) {
-            panic!("Timeout waiting for the old nexus and replica to be removed");
-        }
-        tokio::time::sleep(Duration::from_millis(500)).await;
-    }
-
-    volumes_api.del_volume(&volume.spec.uuid).await.unwrap();
-}
-
-async fn wait_till_replica_disowned(cluster: &Cluster, replica_id: Uuid) {
-    let timeout = Duration::from_secs(RECONCILE_TIMEOUT_SECS);
-    let client = cluster.rest_v00();
-    let specs_api = client.specs_api();
-    let start = std::time::Instant::now();
-    loop {
-        let specs = specs_api.get_specs().await.unwrap();
-        let replica_spec = specs
-            .replicas
-            .into_iter()
-            .find(|r| r.uuid == replica_id)
-            .unwrap();
-
-        if replica_spec.owners.volume.is_none() {
-            return;
-        }
-
-        if std::time::Instant::now() > (start + timeout) {
-            panic!(
-                "Timeout waiting for the replica to be disowned. Actual: '{:#?}'",
-                replica_spec
-            );
-        }
-        tokio::time::sleep(Duration::from_millis(500)).await;
-    }
-}
-
-/// Creates a volume nexus on a node, which will have both spec and state.
-/// Stop/Kill the io-engine container. At some point we will have no nexus state, because the node
-/// is gone. We then restart the node and the volume nexus reconciler will then recreate the nexus!
-/// At this point, we'll have a state again and the volume will be Online!
-async fn missing_nexus_reconcile(cluster: &Cluster) {
-    let volume_client = cluster.grpc_client().volume();
-    let volume = volume_client
-        .create(
-            &CreateVolume {
-                uuid: "1e3cf927-80c2-47a8-adf0-95c486bdd7b7".try_into().unwrap(),
-                size: 5242880,
-                replicas: 1,
-                ..Default::default()
-            },
-            None,
-        )
-        .await
-        .unwrap();
-
-    let rest_api = cluster.rest_v00();
-    let volumes_api = rest_api.volumes_api();
-
-    let volume = volumes_api
-        .put_volume_target(
-            &volume.spec().uuid,
-            cluster.node(0).as_str(),
-            models::VolumeShareProtocol::Nvmf,
-        )
-        .await
-        .unwrap();
-
-    tracing::info!("Volume: {:?}", volume);
-    let volume_state = volume.state;
-    let nexus = volume_state.target.unwrap();
-
-    cluster.composer().stop(nexus.node.as_str()).await.unwrap();
-    let curr_nexus = wait_till_nexus_state(cluster, &nexus.uuid, None).await;
-    assert_eq!(curr_nexus, None);
-
-    cluster.composer().start(nexus.node.as_str()).await.unwrap();
-    let curr_nexus = wait_till_nexus_state(cluster, &nexus.uuid, Some(&nexus)).await;
-    assert_eq!(Some(nexus), curr_nexus);
-
-    let volume_id = volume_state.uuid;
-    let volume = volumes_api.get_volume(&volume_id).await.unwrap();
-    assert_eq!(volume.state.status, models::VolumeStatus::Online);
-
-    volumes_api.del_volume(&volume_id).await.unwrap();
-}
-
-/// Wait until the specified nexus state option matches the requested `state`
-async fn wait_till_nexus_state(
-    cluster: &Cluster,
-    nexus_id: &Uuid,
-    state: Option<&models::Nexus>,
-) -> Option<models::Nexus> {
-    let timeout = Duration::from_secs(RECONCILE_TIMEOUT_SECS);
-    let client = cluster.rest_v00();
-    let nexuses_api = client.nexuses_api();
-    let start = std::time::Instant::now();
-    loop {
-        let nexus = nexuses_api.get_nexus(nexus_id).await;
-        match &nexus {
-            Ok(nexus) => {
-                if let Some(state) = state {
-                    if nexus.protocol != models::Protocol::None
-                        && state.protocol != models::Protocol::None
-                    {
-                        return Some(nexus.clone());
-                    }
-                }
-            }
-            Err(Error::Response(response))
-                if response.status() == StatusCode::NOT_FOUND && state.is_none() =>
-            {
-                return None;
-            }
-            _ => {}
-        };
-
-        if std::time::Instant::now() > (start + timeout) {
-            panic!(
-                "Timeout waiting for the nexus to have state: '{:#?}'. Actual: '{:#?}'",
-                state, nexus
-            );
-        }
-        tokio::time::sleep(Duration::from_millis(500)).await;
-    }
-}
-
-/// Faults a volume nexus replica and waits for it to be replaced with a new one
-async fn hotspare_faulty_children(cluster: &Cluster) {
-    let volume_client = cluster.grpc_client().volume();
-    let registry_client = cluster.grpc_client().registry();
-    let volume = volume_client
-        .create(
-            &CreateVolume {
-                uuid: "1e3cf927-80c2-47a8-adf0-95c486bdd7b7".try_into().unwrap(),
-                size: 5242880,
-                replicas: 2,
-                ..Default::default()
-            },
-            None,
-        )
-        .await
-        .unwrap();
-
-    let volume = volume_client
-        .publish(
-            &PublishVolume::new(volume.spec().uuid.clone(), Some(cluster.node(0)), None),
-            None,
-        )
-        .await
-        .unwrap();
-
-    tracing::info!("Volume: {:?}", volume);
-    let volume_state = volume.state();
-    let nexus = volume_state.target.unwrap().clone();
-
-    let mut rpc_handle = cluster.grpc_handle(cluster.node(0).as_str()).await.unwrap();
-
-    let children_before_fault = volume_children(volume.uuid(), &volume_client).await;
-    tracing::info!("volume children: {:?}", children_before_fault);
-
-    let fault_child = nexus.children.first().unwrap().uri.to_string();
-    rpc_handle
-        .io_engine
-        .fault_nexus_child(FaultNexusChildRequest {
-            uuid: nexus.uuid.to_string(),
-            uri: fault_child.clone(),
-        })
-        .await
-        .unwrap();
-
-    tracing::debug!(
-        "Nexus: {:?}",
-        rpc_handle
-            .io_engine
-            .list_nexus(rpc::io_engine::Null {})
-            .await
-            .unwrap()
-    );
-
-    let children = wait_till_volume_nexus(
-        volume.uuid(),
-        2,
-        &fault_child,
-        &volume_client,
-        &registry_client,
-    )
-    .await;
-    tracing::info!("volume children: {:?}", children);
-
-    assert_eq!(children.len(), 2);
-    // the faulted child should have been replaced!
-    assert!(!children.iter().any(|c| c.uri == fault_child));
-
-    volume_client
-        .destroy(&DestroyVolume::new(volume.uuid()), None)
-        .await
-        .unwrap();
-}
-
-/// Wait for the published volume to have the specified replicas and to not having the specified
-/// child. Wait up to the specified timeout.
-async fn wait_till_volume_nexus(
-    volume: &VolumeId,
-    replicas: usize,
-    no_child: &str,
-    volume_client: &dyn VolumeOperations,
-    registry_client: &dyn RegistryOperations,
-) -> Vec<Child> {
-    let timeout = Duration::from_secs(RECONCILE_TIMEOUT_SECS);
-    let start = std::time::Instant::now();
-    loop {
-        let volume = volume_client
-            .get(GetVolumes::new(volume).filter, None, None)
-            .await
-            .unwrap();
-        let volume_state = volume.entries.clone().first().unwrap().state();
-        let nexus = volume_state.target.clone().unwrap();
-        let specs = registry_client.get_specs(&GetSpecs {}, None).await.unwrap();
-        let nexus_spec = specs.nexuses.first().unwrap().clone();
-
-        if !nexus.contains_child(&ChildUri::from(no_child))
-            && nexus.children.len() == replicas
-            && nexus_spec.children.len() == replicas
-        {
-            return nexus.children;
-        }
-        if std::time::Instant::now() > (start + timeout) {
-            panic!(
-                "Timeout waiting for the volume to reach the specified state (replicas: '{}', no_child: '{}')! Current: {:#?}",
-                replicas, no_child, volume_state
-            );
-        }
-        tokio::time::sleep(Duration::from_secs(1)).await;
-    }
-}
-
-/// Get the children of the specified volume (assumes non ANA)
-async fn volume_children(volume: &VolumeId, client: &dyn VolumeOperations) -> Vec<Child> {
-    let volume = client
-        .get(GetVolumes::new(volume).filter, None, None)
-        .await
-        .unwrap();
-    let volume_state = volume.entries.first().unwrap().state();
-    volume_state.target.unwrap().children
-}
-
-/// Adds a child to the volume nexus (under the control plane) and waits till it gets removed
-async fn hotspare_unknown_children(cluster: &Cluster) {
-    let volume_client = cluster.grpc_client().volume();
-    let registry_client = cluster.grpc_client().registry();
-    let volume = volume_client
-        .create(
-            &CreateVolume {
-                uuid: "1e3cf927-80c2-47a8-adf0-95c486bdd7b7".try_into().unwrap(),
-                size: 5242880,
-                replicas: 2,
-                ..Default::default()
-            },
-            None,
-        )
-        .await
-        .unwrap();
-    let size_mb = volume.spec().size / 1024 / 1024 + 5;
-
-    let volume = volume_client
-        .publish(
-            &PublishVolume::new(volume.spec().uuid.clone(), Some(cluster.node(0)), None),
-            None,
-        )
-        .await
-        .unwrap();
-
-    tracing::info!("Volume: {:?}", volume);
-    let volume_state = volume.state();
-    let nexus = volume_state.target.unwrap().clone();
-
-    let mut rpc_handle = cluster.grpc_handle(cluster.node(0).as_str()).await.unwrap();
-
-    let children_before_fault = volume_children(volume.uuid(), &volume_client).await;
-    tracing::info!("volume children: {:?}", children_before_fault);
-
-    let unknown_replica = format!(
-        "malloc:///xxxxx?size_mb={}&uuid={}",
-        size_mb,
-        ReplicaId::new()
-    );
-
-    // todo: this sometimes fails??
-    // is the reconciler interleaving with the add_child_nexus?
-    rpc_handle
-        .io_engine
-        .add_child_nexus(rpc::io_engine::AddChildNexusRequest {
-            uuid: nexus.uuid.to_string(),
-            uri: unknown_replica.clone(),
-            norebuild: true,
-        })
-        .await
-        .ok();
-
-    tracing::debug!(
-        "Nexus: {:?}",
-        rpc_handle
-            .io_engine
-            .list_nexus(rpc::io_engine::Null {})
-            .await
-            .unwrap()
-    );
-    let children = wait_till_volume_nexus(
-        volume.uuid(),
-        2,
-        &unknown_replica,
-        &volume_client,
-        &registry_client,
-    )
-    .await;
-    tracing::info!("volume children: {:?}", children);
-
-    assert_eq!(children.len(), 2);
-    // the unknown child should have been replaced!
-    assert!(!children.iter().any(|c| c.uri == unknown_replica));
-
-    volume_client
-        .destroy(&DestroyVolume::new(volume.uuid()), None)
-        .await
-        .unwrap();
-}
-
-/// Remove a child from a volume nexus (under the control plane) and waits till it gets added back
-async fn hotspare_missing_children(cluster: &Cluster) {
-    let volume_client = cluster.grpc_client().volume();
-    let registry_client = cluster.grpc_client().registry();
-    let volume = volume_client
-        .create(
-            &CreateVolume {
-                uuid: "1e3cf927-80c2-47a8-adf0-95c486bdd7b7".try_into().unwrap(),
-                size: 5242880,
-                replicas: 2,
-                ..Default::default()
-            },
-            None,
-        )
-        .await
-        .unwrap();
-
-    let volume = volume_client
-        .publish(
-            &PublishVolume::new(volume.spec().uuid.clone(), Some(cluster.node(0)), None),
-            None,
-        )
-        .await
-        .unwrap();
-
-    tracing::info!("Volume: {:?}", volume);
-    let volume_state = volume.state();
-    let nexus = volume_state.target.unwrap().clone();
-
-    let mut rpc_handle = cluster.grpc_handle(cluster.node(0).as_str()).await.unwrap();
-
-    let children_before_fault = volume_children(volume.uuid(), &volume_client).await;
-    tracing::info!("volume children: {:?}", children_before_fault);
-
-    let missing_child = nexus.children.first().unwrap().uri.to_string();
-    rpc_handle
-        .io_engine
-        .remove_child_nexus(rpc::io_engine::RemoveChildNexusRequest {
-            uuid: nexus.uuid.to_string(),
-            uri: missing_child.clone(),
-        })
-        .await
-        .unwrap();
-
-    tracing::debug!(
-        "Nexus: {:?}",
-        rpc_handle
-            .io_engine
-            .list_nexus(rpc::io_engine::Null {})
-            .await
-            .unwrap()
-    );
-    let children = wait_till_volume_nexus(
-        volume.uuid(),
-        2,
-        &missing_child,
-        &volume_client,
-        &registry_client,
-    )
-    .await;
-    tracing::info!("volume children: {:?}", children);
-
-    assert_eq!(children.len(), 2);
-    // the missing child should have been replaced!
-    assert!(!children.iter().any(|c| c.uri == missing_child));
-
-    volume_client
-        .destroy(&DestroyVolume::new(volume.uuid()), None)
-        .await
-        .unwrap();
-}
-
-/// When more than 1 replicas are faulted at the same time, the new replicas should be spread
-/// across the existing pools, and no pool nor any node should be reused
-async fn hotspare_replica_count_spread(cluster: &Cluster) {
-    let nodes = cluster.rest_v00().nodes_api().get_nodes().await.unwrap();
-    assert!(
-        nodes.len() >= 3,
-        "We need enough nodes to be able to add at least 2 replicas"
-    );
-    let pools = cluster.rest_v00().pools_api().get_pools().await.unwrap();
-    assert!(
-        pools.len() >= nodes.len() * 2,
-        "We need at least 2 pools per node to be able to test the failure case"
-    );
-
-    let replica_count = nodes.len();
-    let volume_client = cluster.grpc_client().volume();
-    let registry_client = cluster.grpc_client().registry();
-    let volume = volume_client
-        .create(
-            &CreateVolume {
-                uuid: "1e3cf927-80c2-47a8-adf0-95c486bdd7b7".try_into().unwrap(),
-                size: 5242880,
-                replicas: 1,
-                ..Default::default()
-            },
-            None,
-        )
-        .await
-        .unwrap();
-
-    // stop the core agent, so we can simulate `replica_count-1` data replicas being faulted at once
-    // by increasing the replica count from 1 to `replica_count` under the core agent
-    cluster.composer().stop("core").await.unwrap();
-
-    let mut store = Etcd::new("0.0.0.0:2379")
-        .await
-        .expect("Failed to connect to etcd.");
-
-    let mut volume_spec: VolumeSpec = store.get_obj(&volume.spec().key()).await.unwrap();
-    volume_spec.num_replicas = replica_count as u8;
-    tracing::info!("VolumeSpec: {:?}", volume_spec);
-    store.put_obj(&volume_spec).await.unwrap();
-
-    cluster.restart_core().await;
-
-    let timeout_opts = TimeoutOptions::default()
-        .with_max_retries(10)
-        .with_req_timeout(Duration::from_millis(200))
-        .with_timeout_backoff(Duration::from_millis(50));
-
-    cluster
-        .volume_service_liveness(Some(timeout_opts.clone()))
-        .await
-        .expect("Should have restarted by now");
-
-    // check we have the new replica_count
-    wait_till_volume(
-        volume.uuid(),
-        replica_count,
-        &volume_client,
-        &registry_client,
-    )
-    .await;
-
-    {
-        let volume = cluster
-            .rest_v00()
-            .volumes_api()
-            .get_volume(volume.uuid())
-            .await
-            .unwrap();
-
-        tracing::info!("Replicas: {:?}", volume.state.replica_topology);
-
-        assert_eq!(volume.spec.num_replicas, replica_count as u8);
-        assert_eq!(volume.state.replica_topology.len(), replica_count);
-
-        for node in 0 .. nodes.len() {
-            let node = cluster.node(node as u32);
-            let replicas = volume
-                .state
-                .replica_topology
-                .values()
-                .filter(|r| r.node == Some(node.to_string()));
-            assert_eq!(replicas.count(), 1, "each node should have 1 replica");
-        }
-    }
-
-    volume_client
-        .destroy(&DestroyVolume::new(volume.uuid()), None)
-        .await
-        .unwrap();
-}
-
-/// Remove a replica that belongs to a volume. Another should be created.
-async fn hotspare_replica_count(cluster: &Cluster) {
-    let replica_client = cluster.grpc_client().replica();
-    let volume_client = cluster.grpc_client().volume();
-    let registry_client = cluster.grpc_client().registry();
-    let volume = volume_client
-        .create(
-            &CreateVolume {
-                uuid: "1e3cf927-80c2-47a8-adf0-95c486bdd7b7".try_into().unwrap(),
-                size: 5242880,
-                replicas: 2,
-                ..Default::default()
-            },
-            None,
-        )
-        .await
-        .unwrap();
-
-    let specs = registry_client.get_specs(&GetSpecs {}, None).await.unwrap();
-
-    let replica_spec = specs.replicas.first().cloned().unwrap();
-    let replicas = replica_client
-        .get(GetReplicas::new(&replica_spec.uuid).filter, None)
-        .await
-        .unwrap();
-    let replica = replicas.0.first().unwrap().clone();
-
-    // forcefully destroy a volume replica
-    let mut destroy = DestroyReplica::from(replica);
-    destroy.disowners = ReplicaOwners::from_volume(volume.uuid());
-    match replica_client.destroy(&destroy, None).await {
-        Ok(_) => tracing::info!("replica destroyed forcefully"),
-        Err(_) => tracing::error!("could not destroy replica forcefully"),
-    }
-
-    // check we have 2 replicas
-    wait_till_volume(volume.uuid(), 2, &volume_client, &registry_client).await;
-
-    // now add one extra replica (it should be removed)
-    replica_client
-        .create(
-            &CreateReplica {
-                node: cluster.node(1),
-                name: Default::default(),
-                uuid: ReplicaId::new(),
-                pool: cluster.pool(1, 0),
-                size: volume.spec().size / 1024 / 1024 + 5,
-                thin: false,
-                share: Default::default(),
-                managed: true,
-                owners: ReplicaOwners::from_volume(volume.uuid()),
-            },
-            None,
-        )
-        .await
-        .unwrap();
-
-    wait_till_volume(volume.uuid(), 2, &volume_client, &registry_client).await;
-
-    volume_client
-        .destroy(&DestroyVolume::new(volume.uuid()), None)
-        .await
-        .unwrap();
-}
-
-/// Remove a replica that belongs to a volume. Another should be created.
-async fn hotspare_nexus_replica_count(cluster: &Cluster) {
-    let volume_client = cluster.grpc_client().volume();
-    let registry_client = cluster.grpc_client().registry();
-    let volume = volume_client
-        .create(
-            &CreateVolume {
-                uuid: "1e3cf927-80c2-47a8-adf0-95c486bdd7b7".try_into().unwrap(),
-                size: 5242880,
-                replicas: 2,
-                ..Default::default()
-            },
-            None,
-        )
-        .await
-        .unwrap();
-
-    let volume = volume_client
-        .publish(
-            &PublishVolume::new(volume.spec().uuid.clone(), Some(cluster.node(0)), None),
-            None,
-        )
-        .await
-        .unwrap();
-
-    tracing::info!("Volume: {:?}", volume);
-
-    let timeout_opts = TimeoutOptions::default()
-        .with_max_retries(10)
-        .with_req_timeout(Duration::from_millis(500))
-        .with_timeout_backoff(Duration::from_millis(50));
-    let mut store = Etcd::new("0.0.0.0:2379")
-        .await
-        .expect("Failed to connect to etcd.");
-
-    let mut volume_spec: VolumeSpec = store.get_obj(&volume.spec().key()).await.unwrap();
-    volume_spec.num_replicas += 1;
-    tracing::info!("VolumeSpec: {:?}", volume_spec);
-    store.put_obj(&volume_spec).await.unwrap();
-
-    cluster.restart_core().await;
-
-    cluster
-        .volume_service_liveness(Some(timeout_opts.clone()))
-        .await
-        .expect("Should have restarted by now");
-
-    wait_till_volume_nexus(
-        volume.uuid(),
-        volume_spec.num_replicas as usize,
-        "",
-        &volume_client,
-        &registry_client,
-    )
-    .await;
-
-    let mut volume_spec: VolumeSpec = store.get_obj(&volume.spec().key()).await.unwrap();
-    volume_spec.num_replicas -= 1;
-    tracing::info!("VolumeSpec: {:?}", volume_spec);
-    store.put_obj(&volume_spec).await.unwrap();
-
-    cluster.restart_core().await;
-
-    cluster
-        .volume_service_liveness(Some(timeout_opts.clone()))
-        .await
-        .expect("Should have restarted by now");
-
-    wait_till_volume_nexus(
-        volume.uuid(),
-        volume_spec.num_replicas as usize,
-        "",
-        &volume_client,
-        &registry_client,
-    )
-    .await;
-
-    volume_client
-        .destroy(&DestroyVolume::new(volume.uuid()), None)
-        .await
-        .unwrap();
-}
-
-/// Wait for the unpublished volume to have the specified replica count
-async fn wait_till_volume(
-    volume: &VolumeId,
-    replicas: usize,
-    volume_client: &dyn VolumeOperations,
-    registry_client: &dyn RegistryOperations,
-) {
-    let timeout = Duration::from_secs(RECONCILE_TIMEOUT_SECS);
-    let start = std::time::Instant::now();
-    loop {
-        let specs = registry_client.get_specs(&GetSpecs {}, None).await.unwrap();
-        let replica_specs = specs
-            .replicas
-            .iter()
-            .filter(|r| r.owners.owned_by(volume))
-            .collect::<Vec<_>>();
-
-        if replica_specs.len() == replicas
-            && existing_replicas(volume, volume_client).await == replicas
-        {
-            return;
-        }
-
-        if std::time::Instant::now() > (start + timeout) {
-            panic!(
-                "Timeout waiting for the volume to reach the specified replica ('{}'), current: '{}'",
-                replicas, replica_specs.len()
-            );
-        }
-        tokio::time::sleep(Duration::from_secs(1)).await;
-    }
-}
-
-/// Return the number of replicas that exist and have a state.
-async fn existing_replicas(volume_id: &VolumeId, client: &dyn VolumeOperations) -> usize {
-    let volumes = client
-        .get(GetVolumes::new(volume_id).filter, None, None)
-        .await
-        .unwrap();
-
-    // Get volumes with the given uuid.
-    // There should only be one.
-    let filtered_volumes: Vec<Volume> = volumes
-        .entries
-        .into_iter()
-        .filter(|v| v.uuid() == volume_id)
-        .collect();
-    assert_eq!(filtered_volumes.len(), 1);
-
-    let volume = filtered_volumes[0].clone();
-
-    // A replica is deemed to exist if its topology indicates a valid node and pool.
-    volume
-        .state()
-        .replica_topology
-        .into_iter()
-        .filter(|(_id, topology)| topology.node().is_some() && topology.pool().is_some())
-        .count()
-}
-
-/// Wait for a volume to reach the provided status
-async fn wait_till_volume_status(cluster: &Cluster, volume: &Uuid, status: models::VolumeStatus) {
-    let timeout = Duration::from_secs(RECONCILE_TIMEOUT_SECS);
-    let start = std::time::Instant::now();
-    loop {
-        let volume = cluster.rest_v00().volumes_api().get_volume(volume).await;
-        if volume.as_ref().unwrap().state.status == status {
-            return;
-        }
-
-        if std::time::Instant::now() > (start + timeout) {
-            panic!(
-                "Timeout waiting for the volume to reach the specified status ('{:?}'), current: '{:?}'",
-                status, volume
-            );
-        }
-        tokio::time::sleep(Duration::from_millis(500)).await;
-    }
-}
 
 /// Either fault the local replica, the remote, or set the nexus as having an unclean shutdown
 #[derive(Debug)]
